Regularize affix labelling in “allomorph of” contexts.
There are two possible ways to present an affix: (i) <label> <affix> [<gloss>] and (ii) <affix> <label> [<gloss>]. So far the document is inconsistent. It should instead be consistent, so standardize on (i). Also eliminate affix type labels like “x- prefix” since this information is already present in the representation of the affix. This commit regularizes the labelling in “allomorph of” contexts (easy to search for), but other contexts will need to be adjusted.