Change \vbmorph to use TikZ instead of tabularray
The tabularray package (tblr) is good but it’s extremely slow with large numbers of tblr environments. We switch to using TikZ matrix nodes which also automatically count columns. TikZ is much more flexible for formatting and customization. There’s some blither with \scantokens because TikZ can’t make & an active character, but it’s still much faster than tabularray. The new glossmx style is defined in config/…-tikz.tex but this only applies to the contents of the TikZ matrix; the surrounding tikzpicture environment is in the \vbmorph command which still lives in config/…-macrology.tex.