Add first steps to crossreferencing
First steps to do crossreferencing between morphemes and their entries. The usual LaTeX \label command creates crossreference targets. We add a simple command \X that takes two arguments, an optional one and a mandatory one. The optional argument is the label (with implicit "m:" prefix). The mandatory argument is the text to display. If there is no optional argument then the text is also the label with "m:" prepended to it.

Right now this command is very dumb. It should get smarter and more customizable with time.