Revise morpheme crossreference command
Change to use \Xc command now defined in config/…-colours.tex. The colour can be customized by changing the Xcolor value in that file. The \X command also encloses its argument in \textit to reduce the amount of \fm used. The \X command uses \textit rather than \fm internally because \item[…] in morphlist environments uses \fm which would be undone since \fm is internally \emph and not \textit. In addition, don’t use \X in morpheme strings, particularly the right hand side of a ≡ definition. This is because the morpheme labels for \X differentiate between forms like "abc-" and "abc"; in a morpheme string it is hard to visually separate the morphemes and at the same time include their hyphen/equals. (We could use a tiny space, but that expands the text width.