Change 2nd level bullet in itemize lists
The en-dash is not great as a bullet because it can be confused with the en-dash in front of stems. Move the white bullet ◦ up to second level so it contrasts nicely with the black bullet • at first level. Replace the fourth level with white right-pointing triangle ▷ which, depending on font, should contrast well with the triangular bullet ‣.