Change enumerate refs to include parents
This makes the result of a `\ref{…}` for an enumerate item give a dotted list of the hierarchy of enumerates, reflecting the hierarchical labels already in place. The enumitem package provides automatic hierarchy labels using `label*`, but not automatic hierarchy references. Instead, we have to explicitly refer to the counter for each level of enumerate (`enumi`, `enumii`, `enumiii`, etc.) in every single enumerate level definition. This seems like an ugly hack, but it’s in an official example (`\alph{enumi}.\roman*`) and the manual suggests this is Right Way to do it.